+++ async_input.h
@@ -160,7 +160,7 @@ private:
 	arm_biquad_cascade_df2T_instance_f32 _bufferLPFilter;
 	constexpr static double blockDuration=AUDIO_BLOCK_SAMPLES/AUDIO_SAMPLE_RATE; //[seconds]
 	double _maxLatency=2.*blockDuration; 
-	constexpr static int32_t bufferLength=(int32_t)( 1 + ceil(192000.f/AUDIO_SAMPLE_RATE))*AUDIO_BLOCK_SAMPLES + (int32_t)(ceilf(3.f*blockDuration * 192000.f));	//maximum input sample rate is 192kHz
+	constexpr static int32_t bufferLength=(int32_t)( 1.f + ceilf(192000.f/AUDIO_SAMPLE_RATE))*AUDIO_BLOCK_SAMPLES + (int32_t)(ceilf(3.f*blockDuration * 192000.f));	//maximum input sample rate is 192kHz
 	float _buffer[NOCHANNELS][bufferLength];
 	volatile double _bufferedTime;
 	double _inputFrequency;
@@ -281,7 +281,7 @@ private:
 			_input.setResampleOffset(_resampleOffset);
 			__enable_irq();
 
-			preload(&_bufferLPFilter, diff);
+			preload(&_bufferLPFilter, (float)diff);
 			_resampler.fixStep();		
 		}
 		else {
